backend/requirements.txt
fastapi>=0.110
uvicorn>=0.30
python-multipart
supabase
python-dotenv>=1.0.1
pillow>=9.0.0
numpy>=1.21.0
psycopg[binary]>=3.1.0
timm>=0.9.0
huggingface-hub>=0.19.0
torch>=2.0.0
torchvision>=0.15.0
requests>=2.31.0
httpx>=0.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
